[1] Tweak StatePack Defaults
I changed some settings in the statepack defaults, mostly for convenience when using '-noui- with the Map Editor (especially the audio was quite loud). Alternatively, I could write the changes in the '-noui' argument, but I feel this is cleaner and not a big a problem as the defaults are not used frequently. I also added some constants, assuming XCAM is the "free cam". If you prefer a different constant naming then I can of course change them.