Fixed some invalid mmPhysicsMGR::NewMover assumptions
The NewMover functions seem to have been written under the assumption that if an inst did not have an entity, it could not have been a mover.
However, this is not always true. aiVehicleInstances can be declared as movers before they have attached themselves to an entity.
Also, extended MAX_MOVERS.