Implemented mmPhysicsMGR::TrivialCollideInstances
The original code did not use the bound's Center, even though the Radius is based off that point.
This would cause the calculated distance to be incorrect, breaking collisions with certain props/vehicles.
This was hidden by the usage of CylinderCollisions, which used the banger CG/YRadius values.
Now the original issue is fixed, CylinderCollisions has been disabled.