Disable InvLodFactor
InvLodFactor is used to scale the distance when calculating an instance's LOD.
A value less than 1 increases the max distance at which a LOD is drawn.
However, this value was being calculated completely wrong.
It was inversely proportional to FOV (so a lower "zoomed in" FOV would actually reduce the LOD distances).
Plus it used the horizontal tangent, so would change depending on aspect ratio.