Optimized portals by fixing plane directions.
If we know which way the edge plane is facing relative to the two cells, we can skip visiting edges where the camera and source cell are not on the same side of the plane.
The CHICAGO PTL file has been regenerated using this information.
Saving the PTL also removed any portals which were pointing to missing cells.